requirements.txt
annotated-types==0.7.0
asn1crypto==1.5.1
bcrypt==4.3.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
colorama==0.4.6
coverage==7.4.3
cryptography==44.0.2
factory-boy==3.3.0
Faker==37.1.0
idna==3.10
iniconfig==2.1.0
loguru==0.7.3
lxml==5.2.1
mirakuru==2.6.0
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pluggy==1.5.0
port-for==0.7.4
psutil==7.0.0
psycopg==3.2.6
psycopg-binary==3.2.6
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pytest==8.0.2
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-postgresql==4.1.0
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
reportlab==4.4.0
requests==2.31.0
setuptools==80.0.0
six==1.17.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.4.0
win32_setctime==1.2.0
XlsxWriter==3.2.3
